+++ client/src/main/java/Class222.java
@@ -66,7 +66,7 @@ public final class Class222 {
 		@Pc(6) Class82 local6 = this.aClass82_122;
 		@Pc(16) Class287 local16;
 		synchronized (this.aClass82_122) {
-			local16 = (Class287) this.aClass82_122.method2156((long) arg0);
+			local16 = (Class287) this.aClass82_122.method2156(arg0);
 		}
 		if (local16 != null) {
 			return local16;
@@ -83,7 +83,7 @@ public final class Class222 {
 		}
 		@Pc(66) Class82 local66 = this.aClass82_122;
 		synchronized (this.aClass82_122) {
-			this.aClass82_122.method2150(local16, (long) arg0);
+			this.aClass82_122.method2150(local16, arg0);
 			return local16;
 		}
 	}
